fix-unalign-pointer-in-arm64
For computers using the arm64 of the Apple chip, the link requires strict alignment of pointers in 32-bit form.
In the structure vFlashstr, an 8-bit UINT-sized val is aligned with a (char *) str, resulting in a link error. It is necessary to adjust the consignment of val to 32bit UINT to solve this bug

Resolves ipmitool/ipmitoo#332

+++ include/ipmitool/ipmi_delloem.h
@@ -344,7 +344,7 @@ typedef struct _power_headroom
 } __attribute__ ((packed)) POWER_HEADROOM;
 
 struct vFlashstr {
-	uint8_t val;
+	uint32_t val;
 	const char * str;
 };
 typedef struct ipmi_vFlash_extended_info
